third_party/WebKit/LayoutTests/fast/canvas-api/canvas-skia-excessive-size.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<canvas id="bigCanvas" width="134217728" height="1"></canvas>

<script>
function runTest() {
    var canvas = document.getElementById("bigCanvas");
    var width = canvas.width;
    // We need to perform a context fetch to force allocation of canvas resources.
    var ctx = canvas.getContext("2d");
    assert_false(ctx == null);
    assert_equals(canvas.width, 134217728);
}

async_test(t => {
        window.onload = function() {
            t.step(runTest);
            t.done();
        }
}, 'Verifies if the browser survives the attempted creation of an excessively large canvas.');
</script>
